Adicionando uma ConnectionString de um banco na nuvem(Appharbor) e ajuste no servico de geração de Matricula

+++ ProjetoDeBloco.Dominio/Entidades/Instituicao/Pessoas/Pessoa.cs
@@ -23,10 +23,11 @@ namespace ProjetoDeBloco.Dominio.Entidades.Instituicao.Pessoas
         }
 
         public Guid Id { get; protected set; }
-        public int Matricula { get; set; }
+        public long Matricula { get; set; }
         public string Nome { get; protected set; }
         public DateTime DataNascimento { get; protected set; }
         public bool Ativo { get; protected set; }
+
         public void Ativar()
         {
             if (Ativo == true)
