+++ ProjetoDeBloco.Infraestrutura/Mapeamento/PessoaMap.cs
@@ -17,7 +17,7 @@ namespace ProjetoDeBloco.Infraestrutura.Mapeamento
 
             Property(pessoa => pessoa.Id).HasDatabaseGeneratedOption(DatabaseGeneratedOption.Identity);
             Property(pessoa => pessoa.Nome).HasColumnType("varchar").HasMaxLength(80).IsRequired();
-            Property(pessoa => pessoa.Matricula).HasColumnType("int");
+            Property(pessoa => pessoa.Matricula).HasColumnType("bigint");
             Property(pessoa => pessoa.DataNascimento).HasColumnType("datetime").IsRequired();
             Property(pessoa => pessoa.Ativo).HasColumnType("bit").IsRequired();
         }
